backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.23
pandas==2.2.3
groq==0.11.0
python-dotenv==1.0.1
